+++ z_zone.c
@@ -278,7 +278,7 @@ boolean Z_InitXms(uint32_t size)
 }
 
 
-void Z_MoveConventionalMemoryToExtendedMemory(uint32_t dest, const void __far* src, uint32_t length)
+void Z_MoveConventionalMemoryToExtendedMemory(uint32_t dest, const void __far* src, uint16_t length)
 {
 	ExtMemMoveStruct.Length       = (length + 1) & ~1;
 	ExtMemMoveStruct.SourceHandle = 0;
@@ -289,7 +289,7 @@ void Z_MoveConventionalMemoryToExtendedMemory(uint32_t dest, const void __far* s
 }
 
 
-void Z_MoveExtendedMemoryToConventionalMemory(void __far* dest, uint32_t src, uint32_t length)
+void Z_MoveExtendedMemoryToConventionalMemory(void __far* dest, uint32_t src, uint16_t length)
 {
 	ExtMemMoveStruct.Length       = (length + 1) & ~1;
 	ExtMemMoveStruct.SourceHandle = xmsHandle;
